requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
python-dotenv==0.21.0
requests==2.28.1
spacy==3.4.0
spacy_language_detection==0.2.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1.tar.gz